refactor: Member, PostDetailResponse Converter로 옮기고 Record로 변환
